app/requirements.txt
Flask
openai
python-dotenv
pymongo

